total code split
