+++ web/src/getQueue.ts
@@ -1,0 +1,3 @@
+export const getQueue = () => {
+  return new Array(25).fill("I WILL BE THERE");
+};
