+++ web/src/getResults.ts
@@ -1,0 +1,3 @@
+export const getResults = () => {
+  return new Array(25).fill("EMPTY");
+};
