+++ web/src/lib/iconpaths.ts
@@ -1,0 +1,9 @@
+import { makeIconPath } from "../utils";
+
+export const iconpaths = new Map();
+iconpaths.set("red-tab", makeIconPath("red-search"));
+iconpaths.set("red-tab-icon", makeIconPath("red-search"));
+iconpaths.set("green-tab", makeIconPath("green-search"));
+iconpaths.set("green-tab-icon", makeIconPath("green-search"));
+iconpaths.set("blue-tab", makeIconPath("blue-search"));
+iconpaths.set("blue-tab-icon", makeIconPath("blue-search"));
