+++ web/src/lib/placeholderColors.ts
@@ -1,0 +1,9 @@
+import { bluecolor, darkRed, greencolor } from "../setup";
+
+export const placeholderColors = new Map();
+placeholderColors.set("red-tab", darkRed);
+placeholderColors.set("red-tab-icon", darkRed);
+placeholderColors.set("green-tab", greencolor);
+placeholderColors.set("green-tab-icon", greencolor);
+placeholderColors.set("blue-tab", bluecolor);
+placeholderColors.set("blue-tab-icon", bluecolor);
